+++ R/cpp11.R
@@ -228,6 +228,10 @@ dust_sirs_device_info <- function() {
   .Call(`_dust_dust_sirs_device_info`)
 }
 
+test_cuda_pars <- function(r_device_config, n_particles, n_particles_each, n_state, n_state_full, n_shared_int, n_shared_real, data_size, shared_size) {
+  .Call(`_dust_test_cuda_pars`, r_device_config, n_particles, n_particles_each, n_state, n_state_full, n_shared_int, n_shared_real, data_size, shared_size)
+}
+
 cpp_scale_log_weights <- function(w) {
   .Call(`_dust_cpp_scale_log_weights`, w)
 }