+++ inst/include/dust/interface_helpers.hpp
@@ -4,6 +4,7 @@
 #include <map>
 #include <cpp11/strings.hpp>
 #include <dust/device_info.hpp>
+#include <dust/cuda_launch_control.hpp>
 
 namespace dust {
 namespace interface {
@@ -361,6 +362,42 @@ std::vector<size_t> check_step_snapshot(cpp11::sexp r_step_snapshot,
   }
 
   return step_snapshot;
+}
+
+
+inline dust::cuda::device_config device_config(cpp11::sexp r_device_config) {
+  size_t run_block_size = 128;
+
+  cpp11::sexp r_device_id = r_device_config;
+  if (TYPEOF(r_device_config) == VECSXP) {
+    cpp11::list r_device_config_l = cpp11::as_cpp<cpp11::list>(r_device_config);
+    r_device_id = r_device_config_l["device_id"]; // could error if missing?
+    cpp11::sexp r_run_block_size = r_device_config_l["run_block_size"];
+    if (r_run_block_size != R_NilValue) {
+      int run_block_size_int = cpp11::as_cpp<int>(r_run_block_size);
+      if (run_block_size_int < 0) {
+        cpp11::stop("'run_block_size' must be positive (but was %d)",
+                    run_block_size_int);
+      }
+      if (run_block_size_int % 32 != 0) {
+        cpp11::stop("'run_block_size' must be a multiple of 32 (but was %d)",
+                    run_block_size_int);
+      }
+      run_block_size = run_block_size_int;
+    }
+  }
+  return dust::cuda::device_config(check_device_id(r_device_id),
+                                     run_block_size);
+}
+
+
+inline
+cpp11::sexp device_config_as_sexp(const dust::cuda::device_config& config) {
+  using namespace cpp11::literals;
+  return cpp11::writable::list({"enabled"_nm = config.enabled_,
+                                "device_id"_nm = config.device_id_,
+                                "shared_size"_nm = config.shared_size_,
+                                "run_block_size"_nm = config.run_block_size_});
 }
 
 }